net: shell: Add stats command
Move "stats" command to a separate file.
